refactor: Merged solution on single file
