--- leetcode/1 - twosum/elixir_solution/.formatter.exs
@@ -1,4 +1,0 @@
-# Used by "mix format"
-[
-  inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"]
-]
