--- leetcode/1 - twosum/elixir_solution/test/elixir_solution_test.exs
@@ -1,4 +1,0 @@
-defmodule ElixirSolutionTest do
-  use ExUnit.Case
-  doctest ElixirSolution
-end
