--- leetcode/1 - twosum/elixir_solution/test/test_helper.exs
@@ -1,1 +1,0 @@
-ExUnit.start()
